fix:add tests and separated files
