+++ src/common/interfaces/review-analysis-data.interface.ts
